+++ src/ingest_validation_tools/table_validator.py
@@ -1,7 +1,15 @@
+import csv
+from pathlib import Path
+
 from frictionless import validate as validate_table
 
 
-def get_table_errors(tsv_path, schema):
+def get_table_errors(tsv, schema):
+    tsv_path = Path(tsv)
+    pre_flight_errors = _get_pre_flight_errors(tsv_path, schema=schema)
+    if pre_flight_errors:
+        return pre_flight_errors
+
     report = validate_table(tsv_path, schema=schema,
                             format='csv')
     error_messages = report['errors']
@@ -12,6 +20,42 @@ def get_table_errors(tsv_path, schema):
                 for error in table['errors']
             ]
     return error_messages
+
+
+def _get_pre_flight_errors(tsv_path, schema):
+    try:
+        dialect = csv.Sniffer().sniff(tsv_path.read_text())
+    except csv.Error as e:
+        return [str(e)]
+    delimiter = dialect.delimiter
+    expected_delimiter = '\t'
+    if delimiter != expected_delimiter:
+        return [f'Delimiter is {repr(delimiter)}, rather than expected {repr(expected_delimiter)}']
+
+    # Re-reading the file is ugly, but creating a stream seems gratuitous.
+    with tsv_path.open() as tsv_handle:
+        reader = csv.DictReader(tsv_handle, dialect=dialect)
+        fields = reader.fieldnames
+        expected_fields = [f['name'] for f in schema['fields']]
+        if fields != expected_fields:
+            errors = []
+            fields_set = set(fields)
+            expected_fields_set = set(expected_fields)
+            extra_fields = fields_set - expected_fields_set
+
+            if extra_fields:
+                errors.append(f'Unexpected fields: {extra_fields}')
+            missing_fields = expected_fields_set - fields_set
+            if missing_fields:
+                errors.append(f'Missing fields: {missing_fields}')
+
+            for i_pair in enumerate(zip(fields, expected_fields)):
+                i, (actual, expected) = i_pair
+                if actual != expected:
+                    errors.append(f'In column {i+1}, found "{actual}", expected "{expected}"')
+            return errors
+
+    return None
 
 
 def _get_message(error):
@@ -30,21 +74,24 @@ def _get_message(error):
     On row 2, column "orcid_id", value "bad-id" fails because constraint "pattern" is "fake-re"
 
     '''
-
-    return (
-        f'On row {error["rowPosition"]}, column "{error["fieldName"]}", '
-        f'value "{error["cell"]}" fails because {error["note"]}'
-    )
+    if 'code' in error and error['code'] == 'missing-label':
+        return 'Bug: Should have been caught pre-flight. File an issue.'
+    if 'rowPosition' in error and 'fieldName' in error and 'cell' in error and 'note' in error:
+        return (
+            f'On row {error["rowPosition"]}, column "{error["fieldName"]}", '
+            f'value "{error["cell"]}" fails because {error["note"]}'
+        )
+    return error['message']
 
 
 if __name__ == "__main__":
     import argparse
-    from pathlib import Path
     from yaml import safe_load
 
     parser = argparse.ArgumentParser('CLI just for testing')
-    parser.add_argument('--tsv_path', type=Path, required=True)
-    parser.add_argument('--schema_path', type=Path, required=True)
+    parser.add_argument('--fixture', type=Path, required=True)
     args = parser.parse_args()
-    errors = get_table_errors(args.tsv_path, safe_load(args.schema_path.read_text()))
+    tsv_path = args.fixture / 'input.tsv'
+    schema_path = args.fixture / 'schema.yaml'
+    errors = get_table_errors(tsv_path, safe_load(schema_path.read_text()))
     print('\n'.join(errors))