+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 - Update IMC docs.
 - Add concentration to antibodies.
 - Factor out Frictionless to give us more control over table validation.
+- Check for CSV instead of TSV.
+- Better error message for missing and mis-ordered fields.
 
 ## v0.0.8 - 2021-02-10
 - Update CODEX directory structure